Replace flask-login with basic auth_token management

+++ application/models/Base.py
@@ -8,7 +8,7 @@ class Base(db.Model):
 
     PROTECTED_ATTRIBUTES = ['id', 'date_created', 'date_modified']
 
-    _id = db.Column(db.BigInteger, primary_key=True)
+    _id = db.Column(db.Integer, primary_key=True)
     _date_created = db.Column(db.DateTime, default=db.func.current_timestamp())
     _date_modified = db.Column(
         db.DateTime, default=db.func.current_timestamp(), onupdate=db.func.current_timestamp()
